fix: refactor the package to make it easier to read and to use
