+++ .golangci.yml
@@ -39,7 +39,7 @@ linters:
     - govet
     - ineffassign
     - interfacer
-    - maligned
+    #- maligned
     - misspell
     - nakedret
     - prealloc
